rename+move HumanTask-related APIs
