+++ api/kogito-api-incubation-processes-services/src/main/java/org/kie/kogito/incubation/processes/services/humantask/HumanTaskService.java
@@ -14,18 +14,24 @@
  * limitations under the License.
  */
 
-package org.kie.kogito.incubation.processes.services;
+package org.kie.kogito.incubation.processes.services.humantask;
 
 import org.kie.kogito.incubation.common.DataContext;
 import org.kie.kogito.incubation.common.ExtendedDataContext;
 import org.kie.kogito.incubation.common.LocalId;
+import org.kie.kogito.incubation.common.MetaDataContext;
 
-public interface ProcessAttachmentService {
-    ExtendedDataContext get(LocalId id);
+public interface HumanTaskService {
+    ExtendedDataContext get(LocalId id, MetaDataContext meta);
 
-    ExtendedDataContext add(LocalId taskId, DataContext dataContext);
+    ExtendedDataContext create(LocalId taskId);
 
-    ExtendedDataContext update(LocalId commentId, DataContext dataContext);
+    ExtendedDataContext abort(LocalId taskId, MetaDataContext meta);
 
-    ExtendedDataContext delete(LocalId commentId, DataContext dataContext);
+    ExtendedDataContext complete(LocalId processId, DataContext dataContext);
+
+    ExtendedDataContext update(LocalId processId, DataContext dataContext);
+
+    ExtendedDataContext transition(LocalId processId, DataContext dataContext);
+
 }
